requirements.txt
python-dotenv
requests
pandas
plotly
streamlit
tzdata
openai